Add support for CORS origins defined by regular expressions
Currently, the config is defined in a JSON file. Therefore, all entries to the cors.origins array must be strings. However, the cors npm package used in blueoak-server supports regular expressions as cors origins. In order to utilize that functionality, this PR converts any string in the origin array into a RegExp if it starts with a leading slash.